rag_multiagent/pyproject.toml
[tool.poetry]
name = "rag-multiagent"
version = "0.1.0"
description = ""
authors = ["LoriaAkanksha <akankshaloria21@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sse-starlette = "^2.1.3"
langgraph = "^0.2.19"
langsmith = "0.1.120"
langchain = "^0.2.16"
langchain-groq = "^0.1.10"
langchain-community = "^0.2.16"
arxiv = "^2.1.3"
wikipedia = "^1.4.0"
typing-extensions = "^4.12.2"
python-dotenv = "^1.0.1"
ipython = "^8.27.0"
streamlit = "^1.38.0"
tiktoken = "^0.7.0"
langchainhub = "^0.1.21"
chromadb = "^0.5.5"
langchain-huggingface = "^0.0.3"
cassio = "^0.1.8"
torch = "^2.4.1"
sentence-transformers = "^3.1.0"
fastapi = "^0.114.1"
uvicorn = "^0.30.6"
pydantic = "^2.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
